refactor(issue): move list issues to separate file
